sh2: don't unify shll inputs with u32

+++ tests/end_to_end/sh-dmul/sh2-gcc-o2-out.c
@@ -1,5 +1,5 @@
 s32 test(s32 lhs, s32 rhs) {
-    return (s32) MULT_HI(lhs, rhs);
+    return MULT_HI(lhs, rhs);
 }
 
 u32 test_unsigned(u32 lhs, u32 rhs) {
